% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{draw_canvas}
\alias{draw_canvas}
\title{Create ggplot object with protein chains from feature database}
\usage{
draw_canvas(data = data)
}
\arguments{
\item{data}{Dataframe of one or more rows with the following column
names: 'type', 'description', 'begin', 'end', 'length', 'accession',
'entryName', 'taxid', 'order'. Must contain a minimum of one "CHAIN" as
data$type.}
}
\value{
A ggplot object either in the plot window or as an object.
}
\description{
\code{draw_canvas} uses the dataframe containing the protein features to
creates the basic plot element by determining the length of the longest
protein.
}
\examples{
# draws a blank canvas of the correct size
draw_canvas(five_rel_data)

# combines with draw_chains to plot and label chains.
p <- draw_canvas(five_rel_data)
p <- draw_chains(p, five_rel_data)
p

}
