% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{draw_mainChain}
\alias{draw_mainChain}
\title{Create ggplot2 object with protein `total_length` from feature database}
\usage{
draw_mainChain(p, data = data, label_mainFrame = TRUE, labels = NULL,
  size = 0.5, label_size = 4, margin = 0.1, ...)
}
\arguments{
\item{p}{ggplot2 object ideally created with \code{\link{draw_canvas}}.}

\item{data}{Dataframe of one or more rows with the following column
names: 'type', 'description', 'total_length', 'begin', 'end', 'length', 'accession',
'entryName', 'taxid', 'order'. Must contain a minimum of one "CHAIN" as
data$type.}

\item{label_mainFrame}{Option to label mainFrame or not.}

\item{labels}{Vector with source of names for the mainFrame. EntryName used as
default but can be changed.}

\item{size}{Size of the outline of the mainFrame.}

\item{label_size}{Size of the text used for labels.}

\item{outline}{Colour of the outline of each chain.}

\item{fill}{Colour of the fill of each chain.}
}
\value{
A ggplot2 object either in the plot window or as an object.
}
\description{
\code{draw_mainChain} uses the dataframe containing the protein features to
plot the mainFrame, the full length proteins. It creates the basic plot element
by determining the length of the longest protein. The ggplot2 function
\code{geom_rect} is then used to draw each of the protein
mainFrame proportional to their number of amino acids (length).
}
\examples{
to be supplemented
}
