% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{draw_mol_horiz}
\alias{draw_mol_horiz}
\title{Draw a horizontal domain structure of a protein using base R plot() function}
\usage{
draw_mol_horiz(names, features)
}
\arguments{
\item{names}{List of names e.g. produced by extract_names().}

\item{features}{dataframe of features e.g. produced by extractFeaturesList().}
}
\value{
Plot object in plot window
}
\description{
Draw a horizontal domain structure of a protein using base R plot() function
}
\examples{

}
