% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{draw_motif}
\alias{draw_motif}
\title{Add protein motifs sites to ggplot object.}
\usage{
draw_motif(p, data = data)
}
\arguments{
\item{p}{ggplot object ideally created with \code{\link{draw_chains}}.}

\item{data}{Dataframe of one or more rows with the following column
names: 'type', 'description', 'begin', 'end', 'length', 'accession',
'entryName', 'taxid', 'order'. Must contain a minimum of one "CHAIN" as
data$type.}
}
\value{
A ggplot object either in the plot window or as an object with an
additional geom_rect layer.
}
\description{
\code{draw_motif} adds protein motifs from Uniprot to ggplot object created
by \code{\link{draw_chains}}.
It uses the data object.
The ggplot function
\code{\link[ggplot2]{geom_rect}} is used to draw each of the
motifs proportional to their number of amino acids (length).
}
\examples{
# combines will with draw_domains to plot chains and protein motifs.
p <- draw_chains(five_rel_data, label_size = 1.25)
p <- draw_motif(p, five_rel_data)
p

}
