% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{draw_phosphosites}
\alias{draw_phosphosites}
\title{Draw a horizontal domain structure of a protein using base R plot() function}
\usage{
draw_phosphosites(phospho_loc, radius = 10, colour = "yellow")
}
\arguments{
\item{phospho_loc}{Amino acid location of phosphorylation site}

\item{radius}{Size of circle to designate phosphorylation site}

\item{colour}{Colour of the circle to designate phosphorylation site}
}
\value{
Plot object in plot window
}
\description{
Draw a horizontal domain structure of a protein using base R plot() function
}
\examples{

}
