% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawProteins-package.r
\name{drawProteins}
\alias{drawProteins}
\title{drawProteins.}
\description{
This package has been created to allow the visualisation of protein
schematics based on the data obtained from the
[Uniprot Protein Database](http://www.uniprot.org/).
}
