% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_api.R
\name{extractFeaturesList}
\alias{extractFeaturesList}
\title{Extract features from list into a dataframe}
\usage{
extractFeaturesList(features_list)
}
\arguments{
\item{features_list}{A list of features}
}
\value{
A dataframe of features
}
\description{
Extract features from list into a dataframe
}
\examples{

}
