% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_api.R
\name{extractFeaturesListwithAcc}
\alias{extractFeaturesListwithAcc}
\title{Extract features from JSON object into a dataframe}
\usage{
extractFeaturesListwithAcc(prot_feat)
}
\arguments{
\item{prot_feat}{Original JSON object from GET of Uniprot API}
}
\value{
A dataframe of features
}
\description{
Extract features from JSON object into a dataframe
}
\examples{

}
