% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{five_rel_list}
\alias{five_rel_list}
\title{Features of five human Rel A proteins}
\format{
Large List of 5 elements - one element for each protein
}
\source{
Uniprot {http://www.uniprot.org}
Accession numbers Q04206 Q01201 Q04864 P19838 Q00653
}
\usage{
five_rel_list
}
\value{
Large List of 5 elements - one element for each protein
}
\description{
List of features from five human NFkappaB proteins downloaded from
Uniprot on 15 August 2017
}
\keyword{datasets}
