% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_domains}
\alias{geom_domains}
\title{Add protein domains to ggplot object.}
\usage{
geom_domains(p, label_domains = TRUE, label_size = 4)
}
\arguments{
\item{p}{ggplot object ideally created with \code{\link{geom_chains}}.}

\item{label_domains}{Option to label domains or not.}

\item{label_size}{Size of the text used for labels.}
}
\value{
A ggplot object either in the plot window or as an object with an
additional geom_rect layer.
}
\description{
\code{geom_domains} adds domains to the ggplot object created by
\code{\link{geom_chains}}.
 It uses the prot_data object.
  The ggplot function
 \code{\link[ggplot2]{geom_rect}} is used to draw each of the domain
 chains proportional to their number of amino acids (length).
}
\examples{
# combines will with geom_domains to plot chains and domains.
library(magrittr)
prot_data \%>\%
     geom_chains(label_size = 1.25) \%>\%
     geom_domains(label_size = 1.25) -> p
     p

}
