% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_phospho}
\alias{geom_phospho}
\title{Add protein phosphorylation sites to ggplot object.}
\usage{
geom_phospho(p, size = 2, fill = "yellow")
}
\arguments{
\item{p}{ggplot object ideally created with \code{\link{geom_chains}}.}

\item{size}{Size of the circle}

\item{fill}{Colour of the circle.}
}
\value{
A ggplot object either in the plot window or as an object with an
additional geom_point layer layer.
}
\description{
\code{geom_phospho} adds phosphorylation sites to ggplot object created by
\code{\link{geom_chains}}.
 It uses the prot_data object.
  The ggplot function
 \code{\link[ggplot2]{geom_point}} is used to draw each of the
 phosphorylation sites at their location as determined by prot_data object.
}
\examples{
# combines will with geom_domains to plot chains and phosphorylation sites.
library(magrittr)
prot_data \%>\%
     geom_chains(label_size = 1.25) \%>\%
     geom_phospho() -> p
     p

}
