% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_region}
\alias{geom_region}
\title{Add protein motifs sites to ggplot object.}
\usage{
geom_region(p)
}
\arguments{
\item{p}{ggplot object ideally created with \code{\link{geom_chains}}.}
}
\value{
A ggplot object either in the plot window or as an object with an
additional geom_rect layer.
}
\description{
\code{geom_motif} adds protein motifs from Uniprot to ggplot object created by
\code{\link{geom_chains}}.
 It uses the prot_data object.
  The ggplot function
 \code{\link[ggplot2]{geom_rect}} is used to draw each of the
  motifs proportional to their number of amino acids (length).
}
\examples{
# combines will with geom_domains to plot chains and phosphorylation sites.
library(magrittr)
prot_data \%>\%
     geom_chains(label_size = 1.25) \%>\%
     geom_motif() -> p
     p

}
