% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_repeat}
\alias{geom_repeat}
\title{Add protein repeats sites to ggplot object.}
\usage{
geom_repeat(p, data = data, label_size = 2, outline = "dimgrey",
  fill = "dimgrey", label_repeats = TRUE)
}
\arguments{
\item{p}{ggplot object ideally created with \code{\link{geom_chains}}.}

\item{data}{Dataframe of one or more rows with the following column
names: 'type', 'description', 'begin', 'end', 'length', 'accession',
'entryName', 'taxid', 'order'. Must contain a minimum of one "CHAIN" as
data$type.}

\item{label_size}{Size of text used for labels of protein repeats.}

\item{outline}{Colour of the outline of each repeat.}

\item{fill}{Colour of the fill of each repeat.}

\item{label_repeats}{Option to label repeats or not.}
}
\value{
A ggplot object either in the plot window or as an object with an
additional geom_rect layer.
}
\description{
\code{geom_repeat} adds protein repeats from Uniprot to ggplot object created
by \code{\link{geom_chains}}.
It uses the data object.
The ggplot function
\code{\link[ggplot2]{geom_rect}} is used to draw each of the
motifs proportional to their number of amino acids (length).
}
\examples{
# combines will with geom_domains to plot chains and protein repeat.
p <- geom_chains(five_rel_data, label_size = 1.25)
p <- geom_repeat(p, five_rel_data)
p

}
