% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_gff}
\alias{parse_gff}
\title{Reformat file or url in gff format to allow to draw}
\usage{
parse_gff(file_or_link)
}
\arguments{
\item{file_or_link}{link in gff format or a file in gff format that can be
read by \code{read_tsv} function from the readr package.}
}
\value{
Dataframe of one or more rows with the following column
names: 'accession', 'source', 'type', 'begin', 'end', 'order',
'entryName', 'description'. Must contain a minimum of one "CHAIN" as
data$type to allow plotting.
}
\description{
\code{parse_gff} loads a file or downloads from an url if provided
protein information that is then changed to allow it to work with
\code{draw_canvas} and other draw functions in drawProteins.
}
\examples{


}
