% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_api.R
\name{phospho_site_info}
\alias{phospho_site_info}
\title{Create a dataframe of protein features from JSON object}
\usage{
phospho_site_info(features)
}
\arguments{
\item{features}{A dataframe of protein features, for example created by
the feature_to_dataframe() function.}
}
\value{
A dataframe that only contains protein phosphorylation sites from
Uniprot
}
\description{
Reduces data.frame of features to just phosphorylation sites. Uses a
subsetting step and a grep with the pattern "Phospho".
}
\examples{
sites <- phospho_site_info(prot_data)
head(sites)


}
