% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_json}
\alias{protein_json}
\title{Features of human Rel A protein}
\format{List of 1 with List of 6 inside}
\source{
\url{http://www.uniprot.org/uniprot/Q04206}
}
\usage{
protein_json
}
\description{
List of features from human Rel A downloaded from
Uniprot on 15 August 2017
}
\keyword{datasets}
