% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rel_A_features}
\alias{rel_A_features}
\title{Features of human Rel A protein}
\format{List of 6 - information necessary to draw Rel A/p65}
\source{
\url{http://www.uniprot.org/uniprot/Q04206}
}
\usage{
rel_A_features
}
\description{
List of features from human Rel A downloaded from
Uniprot on 15 August 2017
}
\keyword{datasets}
