% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rel_json}
\alias{rel_json}
\title{Human Rel A protein features in JSON format}
\format{List of 1 with List of 6 - information necessary to draw Rel A/p65}
\source{
\url{http://www.uniprot.org/uniprot/Q04206}
}
\usage{
rel_json
}
\description{
List of 1 with List of 6 inside downloaded from
Uniprot on 1 November 2017
}
\keyword{datasets}
