% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tnfs_data}
\alias{tnfs_data}
\title{Dataframe features of 2 human TNF receptors from Uniprot on 3 Jan 2018}
\format{
A data frame with 127 rows of 9 variables:
\describe{
    \item{type}{type of features - e.g. chain}
    \item{description}{long name for the protein}
    \item{begin}{starting position (amino acid number) of feature}
    \item{end}{ending position (amino acid number) of feature}
    \item{length}{length of feature - number of amino acids}
    \item{accession}{protein Uniprot accession number}
    \item{entryName}{protein Uniprot entry Name}
    \item{taxid}{taxonomic identification - species}
    \item{order}{plotting order from the bottom of the graph}
}
}
\source{
Uniprot {http://www.uniprot.org}
Accession numbers P19438 P25942
}
\usage{
tnfs_data
}
\value{
A data frame with 127 rows and 9 variables
}
\description{
Dataframe features of 2 human TNF receptors from Uniprot on 3 Jan 2018
}
\keyword{datasets}
