% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{draw_canvas}
\alias{draw_canvas}
\title{Create ggplot2 object with protein chains from feature database}
\usage{
draw_canvas(data)
}
\arguments{
\item{data}{Dataframe of one or more rows with the following column
names: 'type', 'description', 'begin', 'end', 'length', 'accession',
'entryName', 'taxid', 'order'. Must contain a minimum of one "CHAIN" as
data$type.}
}
\value{
A ggplot2 object either in the plot window or as an object.
}
\description{
\code{draw_canvas} uses the dataframe containing the protein features to
creates the basic plot element by determining the length of the longest
protein and the number of proteins to plot.
}
\examples{
# draws a blank canvas of the correct size
data("five_rel_data")
draw_canvas(five_rel_data)

}
