% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_api.R
\name{extract_names}
\alias{extract_names}
\title{Extract protein names into a list}
\usage{
extract_names(protein_json)
}
\arguments{
\item{protein_json}{A JSON object}
}
\value{
A list with "accession", "name", "protein.recommendedName.fullName",
gene.name.primary, gene.name.synonym and organism.name.scientific
}
\description{
Converts the list of 6 JSON object created by getting the features from
UniProt. Used in the feature_to_dataframe(). Does not give order. Does not
operate on List of lists - just the list of 6.
}
\examples{

}
